Added texture assets

+++ src/core/platform/unix.cpp
@@ -57,7 +57,7 @@ Platform::Internal::Internal(void)
 void
 Platform::Initialize(void)
 {
-	srand(platform_internal.start_time);
+	srand(static_cast<unsigned int>(platform_internal.start_time));
 }
 
 void
